+++ TractorMod/release-notes.md
@@ -1,5 +1,5 @@
 # Release notes
-## Upcoming release
+## 4.8.1
 * Fixed chests being treated as weeds in some cases.
 * Fixed scythe and some melee weapons not clearing dead crops if enabled.
 * Improved translations. Thanks to Mysti57 (added French)!